requirements.txt
ruamel.yaml
flatdict  # https://pypi.org/project/flatdict/
melddict  # https://pypi.org/project/melddict/

